Move nsfix to a sep package & add desc
Describe why it's needed
